代码复审，更改代码结构
