+++ src/main/java/com/yrmjhtdjxh/punch/domain/Student.java
@@ -1,11 +1,15 @@
 package com.yrmjhtdjxh.punch.domain;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotNull;
 import java.util.Date;
 
 @Data
@@ -21,26 +25,33 @@ public class Student {
     /**
      * 学号
      */
+    @NotNull
     private Long studentID;
 
     /**
      * 密码
      */
+    @NotNull
+    @JsonIgnore
     private String password;
 
     /**
      * 学生姓名
      */
+    @NotNull
     private String name;
 
     /**
      * 性别 1：男 0：女
      */
+    @Max(1)
+    @Min(0)
     private int sex;
 
     /**
      * 学生的年级
      */
+    @NotNull
     private int grade;
 
     /**
@@ -63,8 +74,6 @@ public class Student {
      * 是否在打卡
      */
     private boolean isPunch;
-
-
 
 
 }
