+++ src/main/java/com/yrmjhtdjxh/punch/mapper/PunchRecordMapper.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+@Repository
 public interface PunchRecordMapper {
 
     List<PunchRecord> findByStudentID(Long studentID);
