+++ src/main/java/com/yrmjhtdjxh/punch/service/PunchRecordService.java
@@ -3,14 +3,14 @@ package com.yrmjhtdjxh.punch.service;
 
 import com.yrmjhtdjxh.punch.domain.PunchRecord;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import java.util.Map;
+
 public interface PunchRecordService {
 
-    void insert(PunchRecord punchRecord);
 
-    PunchRecord getUnfinishPunchByStudnetID(Long studentID);
+    Map<String, Object> startPunch(Map<String, String> startPunchMap, HttpSession session, HttpServletRequest request);
 
-    void endRecordTime(PunchRecord punchRecord);
-
-    void deleteShortPunchTime(Long studentID);
-
+    Map<String, Object> endPunch(Map<String, String> startPunchMap, HttpSession session, HttpServletRequest request);
 }
