+++ src/main/java/com/yrmjhtdjxh/punch/service/StudentService.java
@@ -2,11 +2,15 @@ package com.yrmjhtdjxh.punch.service;
 
 import com.yrmjhtdjxh.punch.domain.Student;
 
+import javax.servlet.http.HttpSession;
+import java.util.Map;
+
 public interface StudentService {
     void insert(Student student);
     Student getOne(Long studentID);
     Student checkStudent(Long studentID, String password);
     void changePunch(Long studentID, int punchStatus);
-    void updateStudentInfo(Student student);
-
+    Map<String, String> register(Student student);
+    Map<String, Object> updateStudentInfo(Student student, HttpSession session);
+    Map<String, String> login(Map<String, String> loginMap, HttpSession session);
 }
