+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   datasource:
     username: root
-    password: etron
+    password: "123456"
     url: jdbc:mysql://localhost:3306/punch?useUnicode=true&characterEncoding=utf-8&serverTimezone=Asia/Shanghai
     driver-class-name: com.mysql.cj.jdbc.Driver
     type: com.alibaba.druid.pool.DruidDataSource
@@ -40,7 +40,7 @@ spring:
 
 
 server:
-  port: 8083
+  port: 8085
 
 
 logging:
